+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Ability to list studies in a different project.
   - This was nearly released from Prolific, so now you can specify a `-p` flag to `study list`.
 - Ability to duplicate a study via `study duplicate [id]`.
+- Provide an error message if `PROLIFIC_TOKEN` is not set.
 
 ## 0.0.14